@@ -2,17 +2,19 @@ package passport
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/opentdp/go-helper/strutil"
 
 	"tdp-cloud/cmd/args"
-	"tdp-cloud/helper/strutil"
-	"tdp-cloud/module/model/config"
-	"tdp-cloud/module/model/passport"
-	"tdp-cloud/module/model/user"
+	"tdp-cloud/model/config"
+	"tdp-cloud/model/passport"
+	"tdp-cloud/model/user"
 )
+
+type Controller struct{}
 
 // 注册用户
 
-func register(c *gin.Context) {
+func (*Controller) register(c *gin.Context) {
 
 	var rq *user.CreateParam
 
@@ -22,7 +24,7 @@ func register(c *gin.Context) {
 	}
 
 	// 是否禁止注册
-	if config.ValueOf("registrable") != "true" {
+	if config.ValueOf("Registrable") != "true" {
 		c.Set("Error", "抱歉，已关闭注册功能")
 		return
 	}
@@ -35,7 +37,7 @@ func register(c *gin.Context) {
 
 	rq.Level = 0 //防止逃逸
 	rq.AppKey = strutil.Rand(32)
-	rq.StoreKey = args.Dataset.Secret
+	rq.StoreKey = args.Assets.Secret
 
 	if id, err := user.Create(rq); err == nil {
 		c.Set("Payload", gin.H{"Id": id})
@@ -48,7 +50,7 @@ func register(c *gin.Context) {
 
 // 登录账号
 
-func login(c *gin.Context) {
+func (*Controller) login(c *gin.Context) {
 
 	var rq *passport.LoginParam
 
@@ -71,7 +73,7 @@ func login(c *gin.Context) {
 
 // 获取资料
 
-func profile(c *gin.Context) {
+func (*Controller) profile(c *gin.Context) {
 
 	rq := &user.FetchParam{
 		Id: c.GetUint("UserId"),
@@ -87,7 +89,7 @@ func profile(c *gin.Context) {
 
 // 修改资料
 
-func profileUpdate(c *gin.Context) {
+func (*Controller) profileUpdate(c *gin.Context) {
 
 	var rq *passport.ProfileUpdateParam
 
@@ -107,9 +109,31 @@ func profileUpdate(c *gin.Context) {
 
 }
 
+// 修改头像
+
+func (*Controller) avatarUpdate(c *gin.Context) {
+
+	var rq *passport.AvatarUpdateParam
+
+	if err := c.ShouldBind(&rq); err != nil {
+		c.Set("Error", err)
+		return
+	}
+
+	rq.UserId = c.GetUint("UserId")
+
+	if url, err := passport.AvatarUpdate(rq); err == nil {
+		c.Set("Message", "修改成功")
+		c.Set("Payload", gin.H{"Avatar": url})
+	} else {
+		c.Set("Error", err)
+	}
+
+}
+
 // 统计信息
 
-func summary(c *gin.Context) {
+func (*Controller) summary(c *gin.Context) {
 
 	userId := c.GetUint("UserId")
 	res := passport.Summary(userId)
