@@ -13,8 +13,10 @@ import (
 	"tdp-cloud/api/script"
 	"tdp-cloud/api/taskline"
 	"tdp-cloud/api/terminal"
+	"tdp-cloud/api/upgrade"
 	"tdp-cloud/api/user"
 	"tdp-cloud/api/vendor"
+	"tdp-cloud/api/worker"
 	"tdp-cloud/api/workhub"
 
 	"tdp-cloud/api/alibaba"
@@ -30,7 +32,8 @@ func Router(engine *gin.Engine) {
 
 	api := engine.Group("/api")
 
-	api.Use(midware.OutputHandle())
+	api.Use(midware.OutputHandle)
+	api.Use(midware.JwtGuard)
 
 	{
 		certbot.Router(api)
@@ -42,8 +45,10 @@ func Router(engine *gin.Engine) {
 		passport.Router(api)
 		script.Router(api)
 		taskline.Router(api)
+		upgrade.Router(api)
 		user.Router(api)
 		vendor.Router(api)
+		worker.Router(api)
 		workhub.Router(api)
 
 		alibaba.Router(api)
@@ -55,7 +60,8 @@ func Router(engine *gin.Engine) {
 
 	wsi := engine.Group("/wsi/:auth")
 
-	wsi.Use(midware.SocketHandle())
+	wsi.Use(midware.SocketHandle)
+	wsi.Use(midware.JwtGuard)
 
 	{
 		terminal.Socket(wsi)
