@@ -8,16 +8,18 @@ import (
 
 func Router(api *gin.RouterGroup) {
 
+	ctrl := &Controller{}
+
 	rg := api.Group("/script")
 
-	rg.Use(midware.AuthGuard())
+	rg.Use(midware.AuthGuard)
 
 	{
-		rg.POST("/list", list)
-		rg.POST("/create", create)
-		rg.POST("/detail", detail)
-		rg.POST("/update", update)
-		rg.POST("/delete", delete)
+		rg.POST("/list", ctrl.list)
+		rg.POST("/create", ctrl.create)
+		rg.POST("/detail", ctrl.detail)
+		rg.POST("/update", ctrl.update)
+		rg.POST("/delete", ctrl.delete)
 	}
 
 }
