@@ -4,19 +4,15 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func SocketHandle() gin.HandlerFunc {
+func SocketHandle(c *gin.Context) {
 
-	return func(c *gin.Context) {
+	c.Next()
 
-		c.Next()
+	// 输出错误信息
 
-		// 输出错误信息
-
-		if err, exists := c.Get("Error"); exists && err != nil {
-			c.AbortWithError(exitCode(c, 400), newError(err))
-			return
-		}
-
+	if err, exists := c.Get("Error"); exists && err != nil {
+		c.AbortWithError(exitCode(c, 400), newError(err))
+		return
 	}
 
 }
