@@ -1,8 +1,9 @@
 package migrator
 
 import (
-	"tdp-cloud/module/dborm"
-	"tdp-cloud/module/model"
+	"github.com/opentdp/go-helper/dborm"
+
+	"tdp-cloud/model"
 )
 
 func v100000() error {
@@ -13,7 +14,10 @@ func v100000() error {
 
 func v100000AutoMigrate() error {
 
+	// NOTE: 有外键的表需要先导入
 	return dborm.Db.AutoMigrate(
+		&model.User{},
+		&model.Vendor{},
 		&model.Certjob{},
 		&model.Config{},
 		&model.Cronjob{},
@@ -23,8 +27,6 @@ func v100000AutoMigrate() error {
 		&model.Migration{},
 		&model.Script{},
 		&model.Taskline{},
-		&model.User{},
-		&model.Vendor{},
 	)
 
 }
