@@ -1,16 +1,13 @@
 package migrator
 
 import (
+	"github.com/opentdp/go-helper/strutil"
+
 	"tdp-cloud/cmd/args"
-	"tdp-cloud/helper/strutil"
-	"tdp-cloud/module/model/config"
-	"tdp-cloud/module/model/user"
+	"tdp-cloud/model/user"
 )
 
 func v100001() error {
-
-	// TODO: 临时处理
-	v100000MigrateLog()
 
 	if isMigrated("v100001") {
 		return nil
@@ -31,21 +28,10 @@ func v100001AddUser() error {
 		Password: "123456",
 		Level:    1,
 		AppKey:   strutil.Rand(32),
-		Email:    "cloud@opentdp.org",
-		StoreKey: args.Dataset.Secret,
+		Email:    "user@opentdp.org",
+		StoreKey: args.Assets.Secret,
 	})
 
 	return err
-
-}
-
-func v100000MigrateLog() {
-
-	res, err := config.Fetch(&config.FetchParam{Name: "v100001"})
-
-	if err == nil && res.Id > 0 {
-		config.Delete(&config.DeleteParam{Id: res.Id})
-		addMigration("v100001", "添加默认账号")
-	}
 
 }
