@@ -1,18 +1,28 @@
 package worker
 
 import (
-	"tdp-cloud/helper/logman"
-	"tdp-cloud/helper/psutil"
+	"runtime"
+
+	"github.com/opentdp/go-helper/logman"
+	"github.com/opentdp/go-helper/psutil"
+	"github.com/opentdp/go-helper/socket"
 )
 
-func (pod *RecvPod) Stat(rs *SocketData) error {
+func (pod *RecvPod) Stat(rq *socket.PlainData) error {
 
-	logman.Info("Stat:recv", "taskId", rs.TaskId)
+	var err error
 
-	err := pod.WriteJson(&SocketData{
+	logman.Info("stat:recv", "taskId", rq.TaskId)
+
+	err = pod.WriteJson(&socket.PlainData{
 		Method:  "Stat:resp",
-		TaskId:  rs.TaskId,
-		Payload: psutil.Detail(true),
+		TaskId:  rq.TaskId,
+		Success: true,
+		Payload: map[string]any{
+			"Stat":         psutil.Detail(true),
+			"MemStat":      psutil.GoMemory(),
+			"NumGoroutine": runtime.NumGoroutine(),
+		},
 	})
 
 	return err
