@@ -1,24 +1,28 @@
 package workhub
 
 import (
-	"github.com/mitchellh/mapstructure"
-
-	"tdp-cloud/helper/logman"
-	"tdp-cloud/helper/psutil"
+	"github.com/opentdp/go-helper/logman"
+	"github.com/opentdp/go-helper/psutil"
+	"github.com/opentdp/go-helper/socket"
 )
 
-func (pod *RecvPod) Ping(rq *SocketData) error {
+func (pod *RecvPod) Ping(rq *socket.PlainData) error {
 
-	logman.Info("Ping:recv", "from", pod.Conn.RemoteAddr())
+	var (
+		err  error
+		stat = psutil.SummaryStat{}
+	)
 
-	stat := &psutil.SummaryStat{}
-	if mapstructure.Decode(rq.Payload, stat) == nil {
-		pod.WorkerMeta = stat
+	logman.Info("ping:recv", "from", pod.Conn.RemoteAddr())
+
+	if rq.GetPayload(&stat) == nil {
+		pod.WorkerMeta = &stat
 	}
 
-	err := pod.WriteJson(&SocketData{
+	err = pod.WriteJson(&socket.PlainData{
 		Method:  "Ping:resp",
 		TaskId:  rq.TaskId,
+		Success: true,
 		Payload: "OK",
 	})
 
